implementar paginación en nuestrosAnimales.html

+++ src/main/java/com/grupo3/pawHome/services/UsuarioService.java
@@ -20,7 +20,7 @@ public class UsuarioService {
         return repository.findAll();
     }
 
-    public Optional<Usuario> findById(Long id) {
+    public Optional<Usuario> findById(Integer id) {
         return repository.findById(id);
     }
 }
